refactor: move dataPipeline.ts to enterprise services
Incremental refresh is a premium feature, so move the file gating it
(`validateIncrementalPipeline`) into `enterprise/services/`. While there,
co-locate the two incremental-refresh hash functions (and their
exhaustiveness guard) that previously lived in
`services/experimentTimeSeries.ts` — they're only used by validation and
the incremental refresh query runners, all of which now import from one
enterprise module.

`getFiltersForHash` stays in `experimentTimeSeries.ts` since it's still
shared with the non-incremental hash functions and with
`safeRolloutTimeSeries.ts`; the enterprise file imports it from there.

No behavior change.
