+++ packages/back-end/src/queryRunners/ExperimentIncrementalRefreshQueryRunner.ts
@@ -46,8 +46,8 @@ import {
 import {
   getExperimentSettingsHashForIncrementalRefresh,
   getMetricSettingsHashForIncrementalRefresh,
-} from "back-end/src/services/experimentTimeSeries";
-import { validateIncrementalPipeline } from "back-end/src/services/dataPipeline";
+  validateIncrementalPipeline,
+} from "back-end/src/enterprise/services/dataPipeline";
 import { getExposureQueryEligibleDimensions } from "back-end/src/services/dimensions";
 import { chunkMetrics } from "back-end/src/services/experimentQueries/experimentQueries";
 import { getExperimentById } from "back-end/src/models/ExperimentModel";
