+++ packages/back-end/src/services/experiments.ts
@@ -179,8 +179,8 @@ import { ExperimentIncrementalRefreshExploratoryQueryRunner } from "back-end/src
 import { SourceIntegrationInterface } from "back-end/src/types/Integration";
 import { getExposureQueryEligibleDimensions } from "back-end/src/services/dimensions";
 import { ConcurrentIncrementalRefreshError } from "back-end/src/util/errors";
+import { validateIncrementalPipeline } from "back-end/src/enterprise/services/dataPipeline";
 import { getMetricForSnapshot } from "./reports";
-import { validateIncrementalPipeline } from "./dataPipeline";
 import {
   getIntegrationFromDatasourceId,
   getSourceIntegrationObject,
